+++ client/src/types/Handler/Transporter.ts
@@ -1,4 +1,4 @@
-import { Handler } from 'types';
+import { Handler } from 'types/index';
 
 export interface Transporter extends Handler {
   order: number;
