+++ client/src/types/Handler/index.ts
@@ -1,4 +1,5 @@
 import { Handler } from './Handler';
 import { Site } from './Site';
+import { Transporter } from './Transporter';
 
-export type { Site, Handler };
+export type { Site, Handler, Transporter };
