+++ client/src/types/Manifest/AdditionalInfo.ts
@@ -1,3 +1,6 @@
+/**
+ * Comments and reference information for other objects such Manifest and WasteLine
+ */
 export interface AdditionalInfo {
   originalManifestTrackingNumbers: string[];
   newManifestDestination: NewManifestDestination;
@@ -5,12 +8,18 @@ export interface AdditionalInfo {
   comments: AdditionalInfoComment[];
 }
 
+/**
+ * Comments which can be 'attached' to a handler EPA ID number
+ */
 export interface AdditionalInfoComment {
   label?: string;
   description?: string;
   handlerId?: string;
 }
 
+/**
+ * Used to denote the new destination of rejected waste shipments
+ */
 enum NewManifestDestination {
   OriginalGenerator,
   Tsdf,
