+++ client/src/types/WasteLine/index.ts
@@ -1,15 +1,8 @@
 import {
-  ContainerCode,
   ContainerType,
   QuantityCode,
   QuantityDescription,
   WasteLine,
 } from 'types/WasteLine/WasteLine';
 
-export type {
-  WasteLine,
-  ContainerType,
-  ContainerCode,
-  QuantityCode,
-  QuantityDescription,
-};
+export type { WasteLine, ContainerType, QuantityCode, QuantityDescription };
