--- client/src/types/store/index.ts
@@ -1,3 +1,0 @@
-import { Alert, NotificationState } from 'types/store/notificationSlice';
-
-export type { NotificationState, Alert };
